codebase restructured and test for invalid signature added
