+++ README.md
@@ -1,2 +1,4 @@
 # bbs_sign
 BBS+ Signatures over bn254
+
+Draft: https://identity.foundation/bbs-signature/draft-irtf-cfrg-bbs-signatures.html
