+++ src/key_gen.rs
@@ -4,7 +4,8 @@ use zeroize::{Zeroize, ZeroizeOnDrop};
 use digest::generic_array::{GenericArray, typenum::U48};
 use ark_serialize::{ CanonicalSerialize, CanonicalDeserialize };
 
-use crate::traits_helper::{ hash_to_scalar, FromOkm};
+use crate::utils::utilities_helper::FromOkm;
+use crate::utils::core_utilities::hash_to_scalar;
 
 // Public Key
 #[derive(Debug, Default,CanonicalDeserialize, CanonicalSerialize)]
@@ -18,57 +19,62 @@ pub struct SecretKey{
     pub sk: Fr
 }
 
-// https://identity.foundation/bbs-signature/draft-irtf-cfrg-bbs-signatures.html#name-secret-key
-pub fn key_gen(key_material: &mut [u8], key_info: &[u8],key_dst: &[u8]) -> SecretKey {
+impl SecretKey {
 
-    assert!(key_material.len() >= 32);
-    assert!(key_info.len() <= 65535);
+    // https://identity.foundation/bbs-signature/draft-irtf-cfrg-bbs-signatures.html#name-secret-key
+    pub fn key_gen(key_material: &mut [u8], key_info: &[u8],key_dst: &[u8]) -> Self {
 
-    let mut derive_input = Vec::<u8>::with_capacity(key_material.len() + 2 + key_info.len());
+        assert!(key_material.len() >= 32);
+        assert!(key_info.len() <= 65535);
 
-    derive_input.extend_from_slice(key_material.as_ref());
-    derive_input.extend_from_slice(&[(key_info.len() >> 8) as u8]);
-    derive_input.extend_from_slice(&[(key_info.len() & 0xff) as u8]);
-    derive_input.extend_from_slice(key_info.as_ref());
+        let mut derive_input = Vec::<u8>::with_capacity(key_material.len() + 2 + key_info.len());
 
-    let sk = hash_to_scalar(&derive_input, key_dst);
-    key_material.zeroize();
+        derive_input.extend_from_slice(key_material.as_ref());
+        derive_input.extend_from_slice(&[(key_info.len() >> 8) as u8]);
+        derive_input.extend_from_slice(&[(key_info.len() & 0xff) as u8]);
+        derive_input.extend_from_slice(key_info.as_ref());
 
-    assert!(sk != Fr::from(0));
+        let sk = hash_to_scalar(&derive_input, key_dst);
 
-    SecretKey{
-        sk
+        // zeroize key_material after use
+        key_material.zeroize();
+
+        assert!(sk != Fr::from(0));
+
+        SecretKey{
+            sk
+        }
     }
-}
 
-pub fn sk_to_pk(sk: &SecretKey) -> PublicKey {
-    PublicKey{
-        pk: (G2::generator() * sk.sk).into_affine()}
-}
+    pub fn sk_to_pk(&self) -> PublicKey {
+        PublicKey{
+            pk: (G2::generator() * self.sk).into_affine()}
+    }
 
-// TODO: may not be required. `key_gen` implements the generation of secret key according to the spec
-// https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-bls-signature-05#name-keygen
-// https://github.com/mattrglobal/bbs-signatures/blob/e0ae711ce8da425d671c748201106a5d1bf2bd5b/src/bls12381.rs#L354
-pub fn gen_sk(msg: &[u8]) -> Fr {
-    const SALT: &[u8] = b"BBS-SIG-KEYGEN-SALT-";
-    // copy of `msg` with appended zero byte
-    let mut msg_prime = Vec::<u8>::with_capacity(msg.as_ref().len() + 1);
-    msg_prime.extend_from_slice(msg.as_ref());
-    msg_prime.extend_from_slice(&[0]);
-    // `result` has enough length to hold the output from HKDF expansion
-    let mut result = GenericArray::<u8, U48>::default();
-    assert!(hkdf::Hkdf::<sha2::Sha256>::new(Some(SALT), &msg_prime[..])
-        .expand(&[0, 48], &mut result)
-        .is_ok());
-    let result_array: [u8;48] = result.as_slice().try_into().expect("wrong length!");
-    Fr::from_okm(&result_array)
+    // TODO: may not be required. `key_gen` implements the generation of secret key according to the spec
+    // https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-bls-signature-05#name-keygen
+    // https://github.com/mattrglobal/bbs-signatures/blob/e0ae711ce8da425d671c748201106a5d1bf2bd5b/src/bls12381.rs#L354
+    pub fn gen_sk(msg: &[u8]) -> Self {
+        const SALT: &[u8] = b"BBS-SIG-KEYGEN-SALT-";
+        // copy of `msg` with appended zero byte
+        let mut msg_prime = Vec::<u8>::with_capacity(msg.as_ref().len() + 1);
+        msg_prime.extend_from_slice(msg.as_ref());
+        msg_prime.extend_from_slice(&[0]);
+        // `result` has enough length to hold the output from HKDF expansion
+        let mut result = GenericArray::<u8, U48>::default();
+        assert!(hkdf::Hkdf::<sha2::Sha256>::new(Some(SALT), &msg_prime[..])
+            .expand(&[0, 48], &mut result)
+            .is_ok());
+        let result_array: [u8;48] = result.as_slice().try_into().expect("wrong length!");
+        
+        Self{sk: Fr::from_okm(&result_array)}
+    }
+
 }
 
 #[cfg(test)]
 mod tests {
-
-    use crate::key_gen::key_gen;
-    use crate::key_gen::sk_to_pk;
+    use crate::key_gen::SecretKey;
     use zeroize::Zeroize;
     use ark_bn254::Fr;
 
@@ -78,10 +84,10 @@ mod tests {
         let key_dst = b"BBS-SIG-KEYGEN-SALT-";
 
         // key_info and key_dst are optional
-        let sk1 = key_gen(&mut key_material, &[], key_dst.as_slice());
-        let sk2 = key_gen(&mut key_material, &[], key_dst.as_slice());
-        let pk1 = sk_to_pk(&sk1);
-        let pk2 = sk_to_pk(&sk2);
+        let sk1 = SecretKey::key_gen(&mut key_material, &[], key_dst.as_slice());
+        let sk2 = SecretKey::key_gen(&mut key_material, &[], key_dst.as_slice());
+        let pk1 = SecretKey::sk_to_pk(&sk1);
+        let pk2 = SecretKey::sk_to_pk(&sk2);
 
         // check sk is non-zero
         assert!(sk1.sk != Fr::from(0));
@@ -95,12 +101,12 @@ mod tests {
 
     #[test]
     fn test_zeroize() {
-        let mut key_material = [0u8; 32];
+        let mut key_material = [1u8; 32];
         let key_dst = b"BBS-SIG-KEYGEN-SALT-";
 
         // key_info and key_dst are optional
-        let mut sk = key_gen(&mut key_material, &[], key_dst.as_slice());
-        let _ = sk_to_pk(&sk);
+        let mut sk = SecretKey::key_gen(&mut key_material, &[], key_dst.as_slice());
+        let _ = SecretKey::sk_to_pk(&sk);
         
         // zeroize the secret key after generating public key
         sk.zeroize();
